requirements.txt
pandas
python-dateutil
requests